% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{make_zip_process}
\alias{make_zip_process}
\alias{zip_process}
\title{Create an external zip process}
\usage{
make_zip_process(zipfile, files, recurse = TRUE)

zip_process()
}
\arguments{
\item{zipfile}{Path to the zip file to create.}

\item{files}{List of file to add to the archive. Each specified file
or directory in is created as a top-level entry in the zip archive.}

\item{recurse}{Whether to add the contents of directories recursively.}
}
\value{
A zip process object, its class inherits from the
\link[processx:process]{processx::process} class.
}
\description{
\code{zip_process()} return an R6 class that represents a zip process.
\code{make_zip_process()} is a shorthand that also creates an instance of
this class.
}
