% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{zip_example}
\alias{zip_example}
\title{Get path to zippr example}
\usage{
zip_example()
}
\description{
zippr comes bundled with a sample archive in its \code{inst/extdata} directory.
This function make it easy to access.
}
\examples{
zippr_example()
}
