% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{zip_info}
\alias{zip_info}
\alias{zip_ls}
\title{List information on files in a ZIP archive}
\usage{
zip_info(archive)

zip_ls(archive)
}
\arguments{
\item{archive}{Path to an existing ZIP archive.}
}
\value{
A data.frame with 'fs' class columns.
}
\description{
\code{\link[=zip_info]{zip_info()}} is similar to the \code{unzip -l} command. Compared to
\code{\link[zip:zip_list]{zip::zip_list()}}, the data frame is returned as a tibble when available and
columns use human-readable classes like \code{\link[fs:dir_ls]{fs::dir_info()}}.

\code{\link[=zip_ls]{zip_ls()}} returns only the names of the files in an archive, like
\code{\link[fs:dir_ls]{fs::dir_ls()}}.
}
\details{
USes the \code{options(fs.use_tibble = FALSE)} to disable tibble printing.
}
\examples{
zip_info(zip_example())
zip_ls(zip_example())
}
