% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{zip_size}
\alias{zip_size}
\alias{zip_ratio}
\title{List ZIP archive file sizes and compression ratios}
\usage{
zip_size(archive, sum = FALSE)

zip_ratio(archive, sum = FALSE)
}
\arguments{
\item{archive}{Path to an existing ZIP archive.}

\item{sum}{Whether to return a single size or ratio from the sum of sizes.}
}
\value{
A numeric vector of compression ratios or file sizes.
}
\description{
\code{zip_size()} returns the uncompressed size of files within an archive.

\code{zip_ratio()} returns the division of uncompressed size by compressed size.
}
\examples{
zip_size(zip_example())
zip_ratio(zip_example(), sum = TRUE)
}
