% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zippr.R
\docType{package}
\name{zippr}
\alias{zippr}
\title{zippr: Cross-Platform 'zip' Compression}
\description{
The zippr package is a forked alternative to the \link{zip} package built on
miniz, a cross-platform, lossless, high performance data compression library.
}
\details{
The zippr fork makes three types of changes:
\enumerate{
\item Consistently rename \emph{all} functions with \verb{zip_*} to avoid masking the
\code{utils::zip()} and \code{utils::unzip()} functions from base R.
\item Arrange arguments in a consistent order, accepting the primary file paths
as the first argument for easy piping.
\item Return more data with modern, human readable tibble and fs classes.
}
}
