% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip-package.R
\docType{package}
\name{zip-package}
\alias{zip-package}
\title{zip: Cross-Platform 'zip' Compression}
\description{
Cross-Platform 'zip' Compression Library. A replacement for the 'zip' function, that does not require any additional external tools on any platform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/zip}
  \item \url{https://r-lib.github.io/zip/}
  \item Report bugs at \url{https://github.com/r-lib/zip/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Kuba Podgórski [contributor]
  \item Rich Geldreich [contributor]
  \item Posit Software, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
