% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{zip_list}
\alias{zip_list}
\title{List Files in a 'zip' Archive}
\usage{
zip_list(zipfile)
}
\arguments{
\item{zipfile}{Path to an existing ZIP file.}
}
\value{
A tibble with 'fs' class columns.
}
\description{
Compared to \code{\link[zip:zip_list]{zip::zip_list()}}, archive contents here are given as a tibble
data frame when available (use \code{options(fs.use_tibble = FALSE)} to disable)
and columns use classes like \code{\link[fs:fs_bytes]{fs::fs_bytes()}}, which have numeric values in a
human readable format.
}
\examples{
zip_list(zip_example())
}
\concept{zip/unzip functions}
