% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip_list}
\alias{zip_list}
\title{List Files in a 'zip' Archive}
\usage{
zip_list(zipfile)
}
\arguments{
\item{zipfile}{Path to an existing ZIP file.}
}
\value{
A data frame with columns: \code{filename}, \code{compressed_size},
\code{uncompressed_size}, \code{timestamp}, \code{permissions}.
}
\description{
List Files in a 'zip' Archive

\if{html}{\out{<link rel="stylesheet" type="text/css" href="../doc/assets/extra.css">}}

\if{html}{\out{<script src="../doc/assets/rd.js"></script>}}
}
\concept{zip/unzip functions}
